meters/requirements.txt
paho-mqtt
eth-account
web3
